+++ packages/devon/src/commands/index.ts
@@ -1,5 +1,6 @@
 export * as down from './down';
 export * as exec from './exec';
+export * as genCerts from './gen-certs';
 export * as init from './init';
 export * as install from './install';
 export * as interact from './interact';