+++ packages/devon/src/helpers/index.ts
@@ -1,6 +1,7 @@
 export * from './compose';
 export * from './container';
 export * from './cwd';
+export * from './domains';
 export * from './env';
 export * from './glob';
 export * from './services';