+++ packages/devon/src/templates/.devon.ts
@@ -1,7 +1,8 @@
 export const definitionFile = `import type { DefinitionFile } from '@tnotifier/devon';
 
 export const config: DefinitionFile = {
-    project: '{{project}}',
+    project: '{{project}}',{{#if proxyDomains}}
+    domains: [{{#each proxyDomains}}{{#if @index}}, {{/if}}'{{this}}'{{/each}}],{{/if}}
     services: [{{#each services}}{{#if @index}}, {{/if}}{
         name: '{{this.name}}',{{#if this.path}}
         path: '{{this.path}}',{{/if}}{{#if this.force}}